[Bug 17190] Ensure PCRE is built with NO_RECURSE
By default, regular expression match recursion in PCRE is
performed on the stack, but this means that it's possible to run
out of stack with some combinations of regular expression and
target string.

Defining `NO_RECURSE` during PCRE compilation ensures that PCRE
recurses by allocating space on the heap rather than on the stack.  It
makes matching slower but ensures that regular expression matching is
only limited by the amount of RAM available in the system rather than
the amount of stack space available.