libcairo: Use feature macros to enable Quartz font support APIs
When building Cairo to run on iOS 10, the ApplicationServices
framework and `ATSUFontID` type are not available.  However, the
majority of the Quartz support code works just fine because it
relies purely on the CoreGraphics and CoreText frameworks.

This patch adds three feature macros:

- `CAIRO_QUARTZ_HAS_CORE_GRAPHICS` - define to 1 to indicate
  availability of the CoreGraphics framework.  It should be
  defined on both iOS and MacOS
- `CAIRO_QUARTZ_HAS_APPLICATION_SERVICES` - define to 1 to indicate
  availability of the ApplicationSercives framework.  It should be
  defined only on MacOS
- `CAIRO_QUARTZ_HAS_ATSUFONTID` - define to 1 if the `ATSUFontID`
  type is defined by the ApplicationServices header files.  It
  should normally be defined on MacOS

This patch should be submitted upstream (possibly with some
Autoconf magic).