[Bug 19345] Load CEF from correct location on Linux & Win
When LiveCode is installed by the installer on Linux and Windows, the
CEF library is located in `./Externals/CEF`.  This patch updates the
CEF stubs file to attempt to load `libcef` from the correct location.
This comes at the cost of breaking using the browser widget and
revBrowser from a build tree, where the CEF library is located in
`./CEF` (see bug 19381).