Patch the libffi source to fix an obsolete opcode
The libffi source uses the 'stmeqia' opcode. This is a compound opcode for
the 'stm' (store multiple) instruction with two suffixes: 'eq' to indicate
that it should only execute if the previous comparison result was equality
and 'ia' to indicate the memory layout of the buffer.

The canonical ordering of this suffixes is 'ia' then 'eq' (giving a final
opcode of 'stmiaeq'). Libffi uses the other ordering, which GAS accepts but
Clang's internal assembler does not.

This change simply canonicalises the ordering of the suffixes allowing the
code to compile with either tool.