[[ Build ]] Make use of thirdparty prebuilts optional
This patch modifies the build configuration files of the thirdparty libraries
so that the libraries can either be compiled from source or fetched as prebuilt binaries. This is determined by the 'use_prebuilt_thirdparty' variable which
can be set to a non-zero value to specify using the prebuilt binary files
instead of compiling the libraries from source.