[[ libffi ]] Fix bug in Win32 stdcall closure trampoline
This patch fixes a bug in the Win32 specific closure trampoline
for the stdcall calling convention. Specifically, the `ffi_closure`
context pointer is now fetched from `esp + 4` rather than `esp`.
This is necessary as `FORCEFRAME` is used which pushes a register
onto the stack, meaning that the context pointer pushed by the
generated trampoline is one word higher up.